Use Cuprite
https://github.com/rubycdp/cuprite

Cuprite is a pure Ruby driver (read as no Selenium/WebDriver/ChromeDriver
dependency) for Capybara. It allows you to run Capybara tests on a
headless Chrome or Chromium. Under the hood it uses Ferrum which is
high-level API to the browser by CDP protocol.

Removes capybara_accessibility_audit gem due to incompatibility with
Cuprite.