Use UUIDv7 for primary keys
Enable the pgcrypto PostgreSQL extension.

Ruby 3.3.x supports generating UUIDv7 out of the box; for example:
SecureRandom.uuid_v7

Ruby on Rails will use UUIDv4 by default, this behaviour is changed in
a before_create callback in the ApplicationRecord.

Why use UUIDv7 over UUIDv4?

UUIDv7 includes a UNIX timestamp in the initial bits, solving the
indexing and sorting issues faced by UUIDv4.

Note that this change does increase the odds of collisions and the use
of UUIDv7 may not be appropriate for highly distributed systems.